making the bar service more readable
