+++ RES/src/app/dashboard/components/list/paginated-list/filter-paginated-list/types.interface.ts
@@ -15,3 +15,8 @@ export enum UpdateCallerBarChart {
   BarChartNgSelect,
   SideFilters,
 }
+
+export interface BarComposerHelper {
+  subFlag: boolean;
+  forgetYearsFromStore: boolean;
+}
